- Reimplementing the bytes.unpack instruction, which not only had a   major bug but also was quite inefficient.
- sizes/offsets used with bytes/strings are now 64 bit.

- Fixed bug with undefined behaviour at program startup: the global
  execution context was initalized before the run-time.

- Switching to the most simple GC setup for now:

    * Globally replacing malloc/free at link-time (because the
    selective replacement via hlt_malloc etc. leads to crashes I
    havent' understood yet).

    * Disabling threading support (because I get weird warnings on
    Linux about poor performance).

- Tweaking hilti-build2's optimization:

    * Only one option -O is now supported, not different levels
    O1/2/3. At the beginning of the hilti-build2 script, it's
    hardcoded what this translates to.

    * Right now it translates only into tail-call eliminiation
    because adding any of the -O1/2/3 to clang/opt leads to crashes.
    Not sure why.

+++ hilti/libhilti/str.h
@@ -8,7 +8,7 @@
 #include "exceptions.h"
 #include "rtti.h"
 
-typedef int32_t hlt_string_size;
+typedef int64_t hlt_string_size;
 
 struct hlt_bytes;
